+++ logs/history/info/info-2026-01-03.log
@@ -1,0 +1,50 @@
+2026-01-03 16:41:43.347 INFO  o.h.validator.internal.util.Version      : HV000001: Hibernate Validator 8.0.3.Final
+2026-01-03 16:41:43.412 INFO  piq.piqproject.PiqProjectApplication     : Starting PiqProjectApplication using Java 21.0.4 with PID 15072 (C:\Users\ttkak\OneDrive\Desktop\piqProject\build\classes\java\main started by ttkak in C:\Users\ttkak\OneDrive\Desktop\piqProject)
+2026-01-03 16:41:43.414 INFO  piq.piqproject.PiqProjectApplication     : No active profile set, falling back to 1 default profile: "default"
+2026-01-03 16:41:44.696 INFO  .s.d.r.c.RepositoryConfigurationDelegate : Multiple Spring Data modules found, entering strict repository configuration mode
+2026-01-03 16:41:44.697 INFO  .s.d.r.c.RepositoryConfigurationDelegate : Bootstrapping Spring Data JPA repositories in DEFAULT mode.
+2026-01-03 16:41:44.990 INFO  .s.d.r.c.RepositoryConfigurationDelegate : Finished Spring Data repository scanning in 277 ms. Found 15 JPA repository interfaces.
+2026-01-03 16:41:44.993 INFO  .s.d.r.c.RepositoryConfigurationDelegate : Multiple Spring Data modules found, entering strict repository configuration mode
+2026-01-03 16:41:44.993 INFO  .s.d.r.c.RepositoryConfigurationDelegate : Bootstrapping Spring Data Redis repositories in DEFAULT mode.
+2026-01-03 16:41:45.018 INFO  .s.d.r.c.RepositoryConfigurationDelegate : Finished Spring Data repository scanning in 10 ms. Found 1 Redis repository interface.
+2026-01-03 16:41:47.079 INFO  o.s.b.w.embedded.tomcat.TomcatWebServer  : Tomcat initialized with port 8080 (http)
+2026-01-03 16:41:47.122 INFO  o.a.coyote.http11.Http11NioProtocol      : Initializing ProtocolHandler ["http-nio-8080"]
+2026-01-03 16:41:47.128 INFO  o.apache.catalina.core.StandardService   : Starting service [Tomcat]
+2026-01-03 16:41:47.129 INFO  o.apache.catalina.core.StandardEngine    : Starting Servlet engine: [Apache Tomcat/10.1.44]
+2026-01-03 16:41:47.281 INFO  o.a.c.c.C.[Tomcat].[localhost].[/]       : Initializing Spring embedded WebApplicationContext
+2026-01-03 16:41:47.283 INFO  w.s.c.ServletWebServerApplicationContext : Root WebApplicationContext: initialization completed in 3784 ms
+2026-01-03 16:41:47.772 INFO  o.hibernate.jpa.internal.util.LogHelper  : HHH000204: Processing PersistenceUnitInfo [name: default]
+2026-01-03 16:41:47.862 INFO  org.hibernate.Version                    : HHH000412: Hibernate ORM core version 6.6.26.Final
+2026-01-03 16:41:47.878 INFO  o.h.spatial.integration.SpatialService   : HHH80000001: hibernate-spatial integration enabled : true
+2026-01-03 16:41:47.924 INFO  o.h.c.internal.RegionFactoryInitiator    : HHH000026: Second-level cache disabled
+2026-01-03 16:41:48.112 INFO  com.zaxxer.hikari.HikariDataSource       : HikariPool-1 - Starting...
+2026-01-03 16:41:48.372 INFO  com.zaxxer.hikari.pool.HikariPool        : HikariPool-1 - Added connection org.postgresql.jdbc.PgConnection@3909a854
+2026-01-03 16:41:48.379 INFO  com.zaxxer.hikari.HikariDataSource       : HikariPool-1 - Start completed.
+2026-01-03 16:41:48.425 WARN  org.hibernate.orm.deprecation            : HHH90000026: PostgisPG95Dialect has been deprecated; use org.hibernate.dialect.PostgreSQLDialect instead
+2026-01-03 16:41:48.452 INFO  org.hibernate.orm.connections.pooling    : HHH10001005: Database info:
+	Database JDBC URL [Connecting through datasource 'HikariDataSource (HikariPool-1)']
+	Database driver: undefined/unknown
+	Database version: 15.4
+	Autocommit mode: undefined/unknown
+	Isolation level: undefined/unknown
+	Minimum pool size: undefined/unknown
+	Maximum pool size: undefined/unknown
+2026-01-03 16:41:48.466 INFO  org.hibernate.spatial                    : HHH80000003: hibernate-spatial adding type contributions from : org.hibernate.spatial.dialect.postgis.PostgisDialectContributor
+2026-01-03 16:41:48.718 INFO  o.s.o.j.p.SpringPersistenceUnitInfo      : No LoadTimeWeaver setup: ignoring JPA class transformer
+2026-01-03 16:41:48.744 INFO  org.hibernate.spatial                    : HHH80000003: hibernate-spatial adding type contributions from : org.hibernate.spatial.dialect.postgis.PostgisDialectContributor
+2026-01-03 16:41:49.235 INFO  org.hibernate.spatial                    : HHH80000004: hibernate-spatial adding function contributions from : org.hibernate.spatial.dialect.postgis.PostgisDialectContributor
+2026-01-03 16:41:49.996 INFO  o.h.e.t.j.p.i.JtaPlatformInitiator       : HHH000489: No JTA platform available (set 'hibernate.transaction.jta.platform' to enable JTA platform integration)
+2026-01-03 16:41:50.532 INFO  j.LocalContainerEntityManagerFactoryBean : Initialized JPA EntityManagerFactory for persistence unit 'default'
+2026-01-03 16:41:50.840 INFO  o.s.d.j.r.query.QueryEnhancerFactory     : Hibernate is in classpath; If applicable, HQL parser will be used.
+2026-01-03 16:41:53.438 INFO  p.p.d.users.controller.AdminInitializer  : abcd1234!
+2026-01-03 16:41:54.307 INFO  p.p.d.users.controller.AdminInitializer  : 초기 관리자 계정이 생성되었습니다: admin@example.com
+2026-01-03 16:41:55.064 WARN  JpaBaseConfiguration$JpaWebConfiguration : spring.jpa.open-in-view is enabled by default. Therefore, database queries may be performed during view rendering. Explicitly configure spring.jpa.open-in-view to disable this warning
+2026-01-03 16:41:55.351 INFO  r$InitializeUserDetailsManagerConfigurer : Global AuthenticationManager configured with UserDetailsService bean with name userDetailsServiceImpl
+2026-01-03 16:41:55.870 INFO  o.a.coyote.http11.Http11NioProtocol      : Starting ProtocolHandler ["http-nio-8080"]
+2026-01-03 16:41:55.888 INFO  o.s.b.w.embedded.tomcat.TomcatWebServer  : Tomcat started on port 8080 (http) with context path '/'
+2026-01-03 16:41:55.905 INFO  piq.piqproject.PiqProjectApplication     : Started PiqProjectApplication in 13.155 seconds (process running for 13.907)
+2026-01-03 16:42:15.083 INFO  o.a.c.c.C.[Tomcat].[localhost].[/]       : Initializing Spring DispatcherServlet 'dispatcherServlet'
+2026-01-03 16:42:15.084 INFO  o.s.web.servlet.DispatcherServlet        : Initializing Servlet 'dispatcherServlet'
+2026-01-03 16:42:15.085 INFO  o.s.web.servlet.DispatcherServlet        : Completed initialization in 1 ms
+2026-01-03 16:42:16.341 INFO  p.p.domain.auth.service.AuthService      : 새로운 사용자가 가입되었습니다: email=test6@example.com, nickname=nick6
+2026-01-03 16:42:16.348 INFO  p.p.d.auth.controller.AuthController     : New user registered: test6@example.com
